Finished updating the folder system
