+++ old-files/TopicCardUpdated/ProgressBar.js
