+++ old-files/authorization/authentication-button.js
